j-541django-wagtail-yard/j541-wagtail/djangocode/old/requirements.txt
Django>=2,<3
django-extensions
djangorestframework
django-crispy-forms
django-select2
wagtail